[branch 1.2] Prevent Plugin install if "requires" info is incomplete.
Before install of plugin there is a check to see whether plugin to be installed requires any other plugins as a dependency. If the plugin author has not put a version number against the requires field then only 1 variable enters the array and an error is thrown. This fix checks whether there are at least 2 variables to fill the array ("plugin name", "version no") and if not allocates a 0 for the version number. It also checks if no info (neither plugin name or version no) is given in which case it returns "Unknown Plugin" from the language file.

Note: this second case will not fire at the moment, since the current function skips passed the missing data and takes info from the next line. this may need addressing in the future.

Language file updated to add this change also.